refactor: re-org contract ABIs, misc typings, update imports, code as necessary
